extract LoginUseCase to production code
