+++ lib/features/login/data/login_service/login_service.dart
@@ -1,4 +1,4 @@
-import 'login_result.dart';
+import '../models/login_result.dart';
 
 abstract class LoginService {
   Future<LoginResult> login(String username, String password);
