+++ lib/features/login/data/login_use_case/login_use_case_output.dart
@@ -1,0 +1,6 @@
+import 'package:thepos/features/login/data/models/login_result.dart';
+
+abstract class LoginServiceOutput {
+  void onLoginSuccess(LoginResult result);
+  void onLoginFail(Object error);
+}
