+++ test/login/data/api_login/api_login_service_test.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:thepos/features/login/data/login_service/api_login/api_login_errors.dart';
 import 'package:thepos/features/login/data/login_service/api_login/api_login_service.dart';
-import 'package:thepos/features/login/data/login_service/login_result.dart';
+import 'package:thepos/features/login/data/models/login_result.dart';
 
 import '../../../helpers/mock_client_stub.dart';
 import '../../../invoice/helpers/shared_test_helper.dart';
