+++ test/login/data/helpers/shared_test_helper.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:thepos/features/login/data/login_service/api_login/api_login_errors.dart';
-import 'package:thepos/features/login/data/login_service/login_result.dart';
+import 'package:thepos/features/login/data/models/login_result.dart';
 
 import '../../../invoice/helpers/shared_test_helper.dart';
 import 'api_login_service_sut.dart';
